LayoutTests/http/tests/resourceLoadStatistics/telemetry-generation-expected.txt

Tests that telemetry for prevalent resources is calculated correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Hosts classified as prevalent resources.
FAIL testResult.totalPrevalentResources should be 0 (of type number). Was undefined (of type undefined).
FAIL testResult.totalPrevalentResourcesWithUserInteraction should be 0 (of type number). Was undefined (of type undefined).
FAIL testResult.top3SubframeUnderTopFrameOrigins should be 0 (of type number). Was undefined (of type undefined).
PASS Hosts classified as prevalent resources.
FAIL testResult.totalPrevalentResources should be 4 (of type number). Was undefined (of type undefined).
FAIL testResult.totalPrevalentResourcesWithUserInteraction should be 1 (of type number). Was undefined (of type undefined).
FAIL testResult.top3SubframeUnderTopFrameOrigins should be 4 (of type number). Was undefined (of type undefined).
PASS successfullyParsed is true
Some tests failed.

TEST COMPLETE
